+++ src/main/java/bridge/Application.java
@@ -1,5 +1,8 @@
 package bridge;
 
+import bridge.controller.Controller;
+import bridge.service.BridgeGameService;
+
 public class Application {
 
     public static void main(String[] args) {
